omxil: Keep the old frame height as plane size (nSliceHeight)
If the decoder returns a new size via the crop rect, but hasn't
explicitly indicated a plane height (via the nSliceHeight parameter),
keep the frame height as plane height instead.

Also do the same check for stride vs width. The decoders that had
problems that the nSliceHeight fixes do set nStride properly, so
that part won't be needed there, but might be needed on some
platform.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>