Revert "Notify: add a callback to match telepathy, msn"
The "item-change" callback is called with the playlist lock held in
at least one case: when the now playing meta is cleared while a new
input is created by the playlist.

playlist_CurrentInput() acquires the playlist lock. So the
"item-change" callback cannot call it. This bug is present in the
Growl, MSN and Telepathy modules since version 1.0.0. It got copied
into Notify in version 2.0.0.

This reverts commit fc56b92af1636c53a0545109a74476fe74c054a5.
This should fix #6641.