input: remove BOM skipping hack
This should not be needed (BOM is a zero-width character). If any
stream reader barks on a BOM, it really needs fixing on its own.
The subtitle demuxer was fixed in the previous commit.

In the case of transparent conversion from UTF-16 to UTF-8, the BOM
will be converted to UTF-8 as any other character. Thus the stream
reader can (correctly) detect a UTF-8 text stream.