aout: remove intricate volume/mute transaction system
Volume and mute states are now independant. There is no need to
update them together in a single transaction.

Furthermore, other processes can change the volume and/or mute state
of VLC playback streams asynchronously. Thus volume-up/volume-down
and mute-toggle are not atomic operations even when protected by the
volume lock. We would need to have toggle and up/down provided by the
to the output plugins. That is probably impossible and overkill.
So accept the small race condition and simplify the code.