Remove the nine years old --language hack
This was apparently put in place to work around the lack of language
setting in BeOS. Then it was extended to Windows and MacOS since
gettext failed to retrieve the user language configuration.
BeOS is no longer supported and Windows support and MacOS support were
fixed in gettext 0.18.

Also calling setlocale() from a library can crash.