config: do not save unknown sections
Unknown sections are sections whose name does not match any module in
the module bank. Until now, any unknown section in vlcrc was copied
verbatim when the preferences are saved. In principles, that would
preserve settings for modules that are not available in the current
copy of VLC.

There is a fatal flaw with that approach: When VLC is updated, some
modules may change name. Also some configuration items may be moved
from one module to another. As a consequence, configuration items may
end up in a different section than they previously were. Then vlcrc
will end up with multiple conflicting copies of the same configuration
item: first the current value for the preferences in the new
section, then the old value from the copied old section. When vlcrc is
loaded again, the old copy will override the new one since it is parsed
last. Essentially, the configuration item cannot be changed via
preferences until vlcrc is manually edited or preferences are reset.

I would rather loose settings for old plugins.