contrib: ffmpeg: Use -mincoming-stack-boundary=4
The issue here is that android-x86 doesn't guarantee an aligned stack, which means some of the inline asm functions will waste an extra register on realigning the stack.

libavcodec does however have stuff in place in all entrypoints that realign the stack on all calls into the library. If you build with --extra-cflags='-mincoming-stack-boundary=4', gcc assumes that the stack actually is aligned already, which the entry points ensure, so this inline asm builds fine.

Pointed-out-by: Martin Storsjö <martin@martin.st>

Signed-off-by: Rafaël Carré <funman@videolan.org>