logo: check for negative offsets
if a logo is bigger than a video and if positioning results in negative
offsets, vlc crashes in the blend function.

This patch adds a warning, and offset is forced to 0 if negative.

Note that a bigger logo is not a problem per se. (Blend truncates the logo).
Only offsetting it with negative values was a problem.