playlist: add simple corking infrastructure for audio policy
Depending on the system, this can be called ducking, focus, policy...
But it basically boils down to pausing playback automatically when a
higher priority audio stream starts, typically a voice call.

Resuming playback automatically when communications are over would be
possible as well. But that seems unfriendly to me, so I skipped it.