omxil: Only use pts if it actually is set
If we need to use pts for timestamps, but pts isn't set (the
avi demuxer seems to set it to 0 for cases where it isn't known, while
dts has a nonzero value), use dts instead.

This fixes things on OMAP4 (the only platform where b_use_pts is
set at the moment), when playing back files that have only dts,
no pts.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>