v4l2: copy header (with minor changes) from Linux 3.5
As agreed last week with upstream, that's basicailly the least insane
way to deal with the ever changing and expanding V4L2 API. This allows
compiling VLC with old Linux headers and run it well on a new Linux
kernel. This also avoids a pile of backward compatibility hacks.

A few other userspace tool already copy videodev2.h like that.