fourcc.c: make fourcc lists fit in .rodata
By changing the description from a pointer to an array of characters,
the long tables don't need to be relocated. Unfortunately this adds
30KB of padding zeroes for empty strings.

Since Lookup() returns a custom entry_t object, with a possibly
modified description, this is implemented by creating a new structure
type. With more work it would be possible to simply filling all the
entries with final description and remove the extra matching and entry
manipulation.

Signed-off-by: Diego Elio Pettenò <flameeyes@flameeyes.eu>
Signed-off-by: Rafaël Carré <funman@videolan.org>