Move libvlccore and libcompat to LDFLAGS from LIBADD
An ABI change to libvlccore or libcompat will always be signaled by a
change in a header file. That change will trigger a recompilation and
then relinking of affected plugins.

Thus, there is no need for libvlccore.la and libcompat.la to be
explicit dependencies of plugins. This small hack avoids relinking
every plugin whenever the libvlccore is rebuilt.