Some RTSP implementations (Foscam H.264 IP cameras) return HTTP/1.1 501 Unimplemented on OPTIONS requests. Don't fail out in this case; just continue and assume GET_PARAMETERS is unsupported.
I have a Foscam h.264 IP camera with a very strange RTSP-over-HTTP
server, and I wanted to get VLC to play it back.  Here's the patch
to VLC; I also had a patch committed to live555 to resolve a few
other issues, and that's available in live.2012.09.13.tar.gz.  Do I
need to submit a patch to contrib/src/live555/rules.mak to pick that
new version up?  Or is there a special process for that?  Or should
I just wait a little bit?  :)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>