avcodec: disable threading by default
Threading does not currently work with some codec profiles and is thus
not suited as default. Threading is also known to deadlock in certain
circumstances, such as if no data is ever received (a rather common
scenario with network or DVB reception).

Fixed avcodec version should be checked before this is re-enabled.

Acked-by: Ilkka Ollakka <ileoo@videolan.org>