youtube.lua: use alternative API URL
It seems that the normal URLs included in the web page miss a
"signature" field and return 403 errors now. This approach has a number
of issues, it requires an extra HTTP request, still seems to fail due
to lack of signatures, and have different access restrictions than the
the web pages.

This shall meet the popular demand for now.

Closes #7471