motion: fix orientation of transform filter
The goal of the motion interface is to keep the video upside up. The
angles of the rotate and transform filters are oriented in opposite
directions, yet both filters are treated in the same way by the motion
code: this is incoherent and at least one code path is wrong, regardless
of the orientation of the motion sensors. On my HDAPS T42, the transform
path would turn the video upside down... my take is that its orientation
got flipped somewhere along the way. This fixes that.