Convert leading whitespace characters to non-breaking spaces
If there are any whitespace characters before the start of text in a row,
convert it to non-breaking whitespace. We do this because the RenderHtml
function will strip any cases of more than one contigious whitespace. The
effect of such is that text that is not left justified will be forced to
the left.

Signed-off-by: Rafaël Carré <funman@videolan.org>