Add ability to render the subpicture region, but not the text itself
We want to center the subpicture region, but the text itself needs
to be rendered left justified in order to get consistent placement
of captions (e.g. indentation).

Introduce a new alignment type which centers the subpicture, but
not the text it contains within (and enable it for the EIA-608 decoder).

Signed-off-by: Rafaël Carré <funman@videolan.org>