macosx: fix table column indentation (close #6929)
For indentation, the outline table column needs to be set, which used to be the title column. This is not a bad choice, since it is only column populated by all items.

However, one can't remove the outline table column easily, so I chose to disable the ability to remove it, something virtually all users don't want to do anyway.