Add support for monospace font in HTML renderer
The EIA-608 decoder expects fonts to be rendered with a monospace font,
so add the ability for decoders to make use of the "tt" tag, and tweak
the EIA-608 decoder to use that tag.

This patch supercedes folds in patch 9 from the previous series as well
as incorporates the changes Rafaël Carré solicited from the ML regarding
the fonts to use for non-Linux platforms.

Modified-by: Rafaël Carré <funman@videolan.org>
 - Simplify HandleTT()
 - Fix memleak in Create()
Signed-off-by: Rafaël Carré <funman@videolan.org>