Add support for rendering EIA-608 captions with a black background
Place a black background behind caption text.  We cannot use the
existing "background" feature for subdecoders because it fills the
entire subpicture block, and in the case of the EIA-608 decoder
would result in a large square (since we send all lines to
the subpicture decoder even if they are blank).  This approach
allows for rendering of the background even in such cases, while
preserving proper layout of the captions.

Signed-off-by: Rafaël Carré <funman@videolan.org>