Don't call the legacy Freetype background renderer if the EIA-608 bg render feature is used.
If we're using the new background renderer which just renders the
background under the text (instead of the entire subpicture block),
then don't call the legacy render as well.  This prevents a alpha-
transparent box from being shown when rendering EIA-608 captions
if the user has the Freetype background opacity set to a nonzero
value.

Signed-off-by: Rafaël Carré <funman@videolan.org>