Fix corrupted rendering of EIA-608 captions
When we moved the code to use the legacy method *or* the new method,
we lost the functionality that actually initialized the memory allocated
for the entire subpicture region (including the part where captions
weren't being rendered.

This was most obvious when regression testing on OSX, where it
resulted in old captions continuing to appear on screen even
after being cleared, or garbage in the render buffer.

Add a line that effectively zeros out the buffer prior to use.

Signed-off-by: Rafaël Carré <funman@videolan.org>