ConfigLoadString loses the last character of strings
ConfigLoadString allocates memory for 'size' bytes and copys a NULL terminated
string to this buffer.
ConfigSaveString instead expects the buffer to be NULL terminated but:
 - the 'size' saved to the cache is the char count
 - the string is written without terminating NULL.

So obviously ConfigLoadString does not match the behaviour of ConfigSaveString.
It seems to me that the cached module configuration is not working at all due
to comparing the module file names within CacheFind - this should never have
found a match.

I think this would need some more testing.

Regards,
Mario

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>