libmp4: small bug fix
When MP4_ReadBoxContainerChildren() returns, one expects the position in
the stream to be that of the next box after the last parsed box.

Without this very patch, if the last parsed box is an "unknown" one,
the position in the stream would be at the begining of that box,
preventing playback of latest Anevia's MS Smooth Streaming live streams,
for example.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>