dbus: Standardise property getters
There was a lot of repeated code, some of which properly checked return
values and some of which didn't.  This is all replaced and folded into
the GetProperty function of each interface.

A side-effect is that it should be more efficient (we only check the
interface argument once instead of once for each property, and there
should be a smaller overall code size).

Signed-off-by: Mirsal Ennaime <mirsal@videolan.org>