omxil: Pass extradata for WMV3 according to OMX specs
Some OMX decoders that can handle WMV3 (such as the broadcom one
in raspberry pi) can handle the WMV3 extradata both in the original
form as in ASF, and in this format, while the one in Samsung Galaxy
S3 requires it to be in this format (as the OMX IL 1.2 specs specify).

This makes HW-accelerated WMV3 decoding work on Galaxy S3.

The code can be split out to a reusable function if/when the same
needs to be done elsewhere.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>