omxil: Ignore all "secure"/DRM codecs
We can't use any of them since they don't output plain YUV data
but require using direct rendering.

This simplifies the condition and possibly also catches other cases
where it would be needed.

The case where it was added was for a device where the codec ending
with .secure was listed before the normal one. This kind of codec
is available on many other devices as well, but they're normally listed
only after the normal ones so we never pick them.

Signed-off-by: Martin Storsjö <martin@martin.st>