omxil: Add an omxil based video output module
This is kept under modules/codes/omxil due to sharing much code
with the omxil codec.

This is only tested with the OMX.broadcom.video_render component
on the Raspberry Pi so far, which in itself doesn't advertise any
roles at all, so this could possibly use the "iv_renderer.yuv.overlay"
role instead of plan "iv_renderer".

But it should be possible to generalize this further and adapt it to
any OMX IL video renderer component.

Signed-off-by: Martin Storsjö <martin@martin.st>