omxil: Use the alternative 64 bit integer ABI on the raspberry pi
This changes how some structs are laid out.

Most of OMX API implementation on the raspberry pi runs on the
VideoCore processor and not on the normal ARM thar runs the
userland. Some OMX structs are passed over to this processor as
whole blocks, so the OMX ABI used in the userland is tied to be the
same as the one used by the VideoCore firmware.

Signed-off-by: Martin Storsjö <martin@martin.st>