aout: remove what should be an unneeded update of "mute"
if aout_MuteSet returns successfully, the "mute" variables at the aout
and playlist levels have already been updated via a call to aout_MuteReport.

This late update was a hack to make things work when a audio output driver
forgets to call aout_MuteReport. So this patch may bring regression for those
audio output drivers that still forget to call aout_MuteReport and yet
terminate successfully. This is likely to occur in corner cases, for instance
when no stream is available as was the case for pulse.c