equalizer: increase precision of frequency centers in debug output
Currently, one of the debug lines in the equalizer module prints out center
frequency values truncated (not rounded) to zero decimal places, and when VLC
is set to use ISO center frequencies, this causes some level of disagreement
between the depiction of the truncated frequencies in the module debug output
and the depiction of the rounded frequencies in the Qt GUI labels.  The 62.5 Hz
ISO center frequency is one of these frequencies, where we have 62 Hz shown in
the debug output and 63 Hz shown in the Qt GUI.  This patch fixes the issue by
displaying the center frequencies in the debug output as real values rounded to
the nearest hundredth.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>