avcodec: Compile again with FFmpeg libavcodec library.
Make fails when using FFmpeg libavcodec version newer than checked by
LIBAVCODEC_VERSION_CHECK, because of redeclaration of many enumerators,
that are unnecessary with newer version of libavcodec library.
It is due to missing parenthesis in macro definition that are making
it wrong.

Probably same with libav libavcodec library, too.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>