skins2: fix video not properly tied to the right video control
Certain skins like wmp12 may fail to deal with video controls in some
corner cases. For this skin, no window was set visible and the skin
engine eventually fell back to force the first window to become visible.

This patch is expected to fix trac #8368 (need to be tested on windows)