equalizer: Enforce type correctness for M_PI as well
It was my expectation that M_PI in the EqzCoeffs function is automatically cast
to a float during compile time, but my expectation turned out to be incorrect.
Specifically, I noticed in GCC's assembly output of equalizer.c that GCC was
doing the inverse by making the program convert all single-precision terms
(excluding 2.0f * M_PI) in the line containing M_PI to double-precision, and
then making it convert the double-precision result to single-precision before
the assignment to f_theta_1.  As a result, M_PI must be explicitly cast to a
float.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>