x11: do not pretend that the video window was resized (refs #8696)
If the core requests a change of the size of the video window, and
the display plugin returns success, the display configuration stored in
the core goes out of synchronization with the X11 server. This results
in incorrect video placement.

Since it cannot be known if resize will succeed or not (at least not
without patching all window provider plugins to block and cross-check),
just return an error.