Set up the encoder after the transcode filter chain
The filter chain can change the format (in particular by changing
colorspace, or by deinterlacing to twice the frame rate), so wait with
initializing the encoder format until the filter chain has been set up.

It might be that the encoder itself demands format changes, e.g. due to
chroma demands or to make the resolution divisible by 16, but we can add
that extra filter to the end of the filter chain after the encoder
parameters have been set up.

Also fix a bug where we'd leak the filter chain when changing it during
the encode.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>