avcodec: mark get_buffer thread safe
In libavcodec, thread-safe callback merely means callback that can run
on any thread. It does not imply that the codec should be re-entrant.
In other words, it boils down to not using thread-specific variables.

The VLC get_buffer callback is not reentrant currently, but it does not
care which thread it runs on.