skins2: improve fullscreen behaviour
At the core level, the 'f' hotkey and the double click are managed in a
slightly different way. The former sets up both the playlist and the vout
"fullscreen" variables whereas the latter only sets up the vout variable.

This patch first relies on the vout state if any, and falls back to the
playlist state. This ensures that the fullscreen button always works
consistently for the user. Otherwise, user needs to press twice instead
of once when the playlist and the vout are no longer in sync (after a double
click)