vout: avoid picture copy when decoder and output format do not match
When the decoder output and the display video formats are different,
the picture copy comes for "free" during the conversion from the
earlier format to the latter one. So do not bother with copying the
picture before the conversion.

The straight copy is still required however in the case that the video
formats are identical but the display pool is inadequate (too few
pictures, slow memory bus...).

Signed-off-by: Rafaël Carré <funman@videolan.org>