contribs: Add a patch for libpng for building with automake 1.11
automake 1.12 changed the default for tests, requiring projects
that need serial test execution to add an option for setting this.
This option isn't available on automake 1.11, so those projects
became incompatible with automake 1.11 once they got ready for 1.12.
This obviously isn't an issue in most cases where the prebuilt
configure is good enough.

Since we don't run the tests in contribs, just remove this option
for now, to allow building with both automake 1.11 and 1.12.

Signed-off-by: Martin Storsjö <martin@martin.st>