omxil: Add a new qualcomm specific custom pixel format
This is pretty much normal NV12, but with stride and plane
heights aligned to 32 (according to a commit on code aurora
forum). The actual values are signalled in the normal stride
and slice height fields, so therefore it can be treated as
normal NV12 as far as our code is concerned.

This is returned on the Sony Xperia Z1 (Snapdragon 800) -
mediacodec works (with a quirk mentioned in the previous commit),
while our omxil/iomx code fails to allocate buffers.

Signed-off-by: Martin Storsjö <martin@martin.st>