androidsurface: Only use ANativeWindow if the private symbols aren't found
Even if ANativeWindow is a public API, the release call seems to
crash on a Nexus One (running 2.3) and on a Galaxy Tab (running 3.2).
The exact reason is not known or understood yet, but it might be
due to accessing and dealing with the Surface from both Java (via the
SurfaceHolder class) and via the ANativeWindow API.

Therefore, only use the ANativeWindow if the old methods aren't
found (that is, on 4.3).

The ANativeWindow output works fine on firmwares as early as 4.0
though.

Signed-off-by: Martin Storsjö <martin@martin.st>