avformat mux: Propagate seekable status into avformat.
Some muxes, in particular mkv/webm, behave very differently depending on
whether we say that the stream is seekable or not (by providing the IOSeek
function). It does not help that the seek function itself returns an error.

Thus, add a new access_out control called ACCESS_OUT_CAN_SEEK, set to true
for seekable files in the file output only, and propagate the status of that
into avformat at initialization time.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>