Qt: do not free module config too early
Fix regression introduced in 95f145933a553e082eb42577d72dbf8bdea9972b

ConfigControl::getName() uses memory of p_item which is a part of p_config memory. This was causing a crash when trying to save the preferences, as the constructor had already destroyed it upon creation.

Instead, free the p_config memory on the destructor, which is automatically called when the dialog closes, saved or cancelled. Verified to not leak any memory.

Close #9214

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>