PulseAudio: remove broken logic
sys->first_pts is the time the stream started originally, but was
mistaken as the time of the stream was paused.

Now, it is assumed that pausing and resuming take approximately the
same time. This is usually true, and if it is not, there is no way for
either VLC or PulseAudio to know (since ALSA does not expose
pause/resume latency informations for the underlying devices).

In practice, this makes no real difference other than removing a bogus
warning about being late when resuming from pause.