access: remove recursion
If you missed the fact that recursive makefiles are harmful, there is
plenty of documentation online (try "Recursive makefiles considered
harmful"), or ask the libav developers for the more concise version.

This change and the previous ones, are particularly boring though. The
main motivation is to eventually address #9397, after the same change is
applied to codec/, demux/ and possibly a few other subdirectories.

Note that some more simplification between Makefile.am files and
configure.ac is definitely possible - if someone is bored.