LGPL relicensing of streaming output
Re-license almost all the streaming modules to LGPLv2.1+ with
authorization from their respective contributors

After access_output, this includes stream_out and mux.

The duplicate and rtp modules are not concerned by this relicensing
because of unreachable authors. Notably:
 - An L. Ber (dedd7cd2)
 - Andy chenee (5caa2ba)

Thanks to everyone for their help and answers in due time